Changing label types from int64 to int32. (#91)
For classification datasets that already have unit tests, the type of label series is changed from default (int64) to int32 (that is expected by most ML libraries). For datasets that do not have unit tests yet, this needs to be done in future commits.

+++ training/xtime/datasets/_eye_movements.py
@@ -77,7 +77,7 @@ class EyeMovementsBuilder(DatasetBuilder):
         assert isinstance(y, pd.Series), f"Expecting y to be of type pd.Series (type = {type(y)})."
 
         # Encode labels. Move from `category` type to int type with labels [0, 1, 2]
-        y = y.astype(int)
+        y = y.astype("int32")
 
         #
         _binary_features = ["P1stFixation", "P2stFixation", "nextWordRegress"]
